Put the displaying logic in Game.java
We can switch it to some Algorithmic Classes later on,
but in common sense it is good to distribute the work in the main class.

Here I parse the given Message an give some example Answers.

Deleted the GameMessageProcessor.java because this class couldn't do much, without complete knowledge of the Game.java
In Effekt I let Game.java implement InputHandler.java.

Last change. I developed a Filedescriptor System, to addresse a number of Textareas in the Console.java.

Therefore we have a Simple println, for printing on stdout (the Textarea you put in form the start) and anotherone, to address a specific Area. Gave an enum for eventually used areas (Market for spaceship, planets for both)

But I am not satisfied with this solution. For this to work a Game class has to know to mush about a Consol.